+++ backend/canisters/user/impl/src/updates/c2c_install_bot.rs
@@ -35,10 +35,13 @@ fn c2c_install_bot_impl(args: Args, state: &mut RuntimeState) -> OCResult {
     }
 
     // If there isn't already a direct chat with the bot, create one now
-    let chat = state
-        .data
-        .direct_chats
-        .get_or_create(args.bot_id, UserType::BotV2, || state.env.rng().random(), now);
+    let chat = state.data.direct_chats.get_or_create(
+        state.env.canister_id().into(),
+        args.bot_id,
+        UserType::BotV2,
+        || state.env.rng().random(),
+        now,
+    );
 
     // Subscribe to permitted chat events
     if let (Some(subscriptions), Some(permissions)) = (args.default_subscriptions, args.granted_autonomous_permissions.clone())