+++ backend/integration_tests/src/communities/delete_channel_tests.rs
@@ -117,9 +117,10 @@ fn stable_memory_garbage_collected_after_deleting_channel() {
         get_stable_memory_map(env, community_id, STABLE_MEMORY_MAP_MEMORY_ID).len(),
         initial_stable_memory_map_keys + 100
     );
+    // A message id and an expiring event for each message, plus the sender's metrics
     assert_eq!(
         get_stable_memory_map(env, community_id, STABLE_MEMORY_MAP_SMALL_ENTRIES_MEMORY_ID).len(),
-        initial_small_entries_keys + 200
+        initial_small_entries_keys + 201
     );
 
     for _ in 0..80 {
@@ -132,7 +133,7 @@ fn stable_memory_garbage_collected_after_deleting_channel() {
     );
     assert_eq!(
         get_stable_memory_map(env, community_id, STABLE_MEMORY_MAP_SMALL_ENTRIES_MEMORY_ID).len(),
-        initial_small_entries_keys + 280
+        initial_small_entries_keys + 282
     );
 
     client::community::happy_path::delete_channel(env, user1.principal, community_id, channel_id1);
@@ -146,7 +147,7 @@ fn stable_memory_garbage_collected_after_deleting_channel() {
     );
     assert_eq!(
         get_stable_memory_map(env, community_id, STABLE_MEMORY_MAP_SMALL_ENTRIES_MEMORY_ID).len(),
-        initial_small_entries_keys + 80
+        initial_small_entries_keys + 81
     );
 
     client::community::happy_path::delete_channel(env, user1.principal, community_id, channel_id2);